refactor: Consolidate sso validation code in to new file
This moves+updates existing sso validation code in
to its own module so that it can be reused elsewhere.

The new file created is `src/auth/sso/validation.ts`

Additionally, the existing code that uses this validation
is updated to use the new changes.
